Fix signed/unsigned warning in bn_test.cc.
BN_num_bits returns unsigned.

Reviewed-on: https://boringssl-review.googlesource.com/4895

+++ crypto/bn/bn_test.cc
@@ -1351,14 +1351,15 @@ static bool test_mod_sqrt(FILE *fp, BN_CTX *ctx) {
 }
 
 static bool test_small_prime(FILE *fp, BN_CTX *ctx) {
-  static const int kBits = 10;
+  static const unsigned kBits = 10;
 
   ScopedBIGNUM r(BN_new());
-  if (!r || !BN_generate_prime_ex(r.get(), kBits, 0, NULL, NULL, NULL)) {
+  if (!r || !BN_generate_prime_ex(r.get(), static_cast<int>(kBits), 0, NULL,
+                                  NULL, NULL)) {
     return false;
   }
   if (BN_num_bits(r.get()) != kBits) {
-    fprintf(fp, "Expected %d bit prime, got %d bit number\n", kBits,
+    fprintf(fp, "Expected %u bit prime, got %u bit number\n", kBits,
             BN_num_bits(r.get()));
     return false;
   }
